requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.2
chromadb==0.4.7
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.2.0
dataclasses-json==0.5.14
decorator==5.1.1
distro==1.8.0
docstring_parser==0.16
docx==0.2.4
emoji==2.8.0
executing==2.0.1
fastapi==0.99.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.1
langchain-chroma==0.1.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-google-vertexai==1.0.4
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langsmith==0.1.63
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
onnxruntime==1.16.3
openai==1.30.3
orjson==3.10.3
overrides==7.4.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
posthog==3.0.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.13
pydantic_core==2.14.5
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.23.1
PyMuPDFb==1.23.0
pyparsing==3.1.2
pypdf==3.16.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
streamlit==1.35.0
streamlit-chat==0.1.1
streamlit-searchbox==0.1.12
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.1
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.2
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
unstructured==0.10.8
uritemplate==4.1.1
urllib3==2.1.0
utils==1.0.2
uvicorn==0.24.0.post1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.3
